Add ability to offset GCode program using probed surface height.

+++ src/machine/Machine.h
@@ -33,17 +33,17 @@ struct JobProgress {
     GLine executing_line;
 
     // Stats
-    float running_time_s;
-    float distance_moved;
+    double running_time_s;
+    double distance_moved;
 };
 
 struct ProbeResult {
     EIGEN_MAKE_ALIGNED_OPERATOR_NEW
     bool contact_made;
 
-    Eigen::Vector3f probe_start;
-    Eigen::Vector3f probe_direction;
-    Eigen::Vector3f contact_point;
+    Eigen::Vector3d probe_start;
+    Eigen::Vector3d probe_direction;
+    Eigen::Vector3d contact_point;
 };
 
 // Pure Virtual Interface
@@ -61,7 +61,7 @@ public:
     // returns a shared progress object which can be used to query the programs state
     virtual const std::shared_ptr<JobProgress> QueueProgram(const GProgram& program) = 0;
 
-    virtual std::future<ProbeResult> ProbeSurface(const Eigen::Vector3f& probe_direction, float feed_rate) = 0;
+    virtual std::future<ProbeResult> ProbeSurface(const Eigen::Vector3d& probe_direction, double feed_rate) = 0;
 
     //        virtual void Pause() = 0;
     //        virtual void Continue() = 0;
